part4/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
marshmallow==3.21.1
mysqlclient==2.2.4
python-dotenv==1.0.1